App: fix DB table fourigen key type

+++ database/migrations/2019_07_18_021317_create_designers_table.php
@@ -22,7 +22,7 @@ class CreateDesignersTable extends Migration
             $table->string('country_code', 10);
             $table->string('phone')->nullable();
             $table->string('portfolio_name', 25)->nullable();
-            $table->bigInteger('gender_id')->default(null);
+            $table->unsignedInteger('gender_id')->default(null);
             $table->boolean('is_active')->default(0);
             $table->boolean('is_verified')->default(0);
             $table->string('verification_code', 1000)->default(null);
