+++ database/migrations/2019_07_18_021752_create_admins_table.php
@@ -21,7 +21,7 @@ class CreateAdminsTable extends Migration
             $table->string('email')->unique();
             $table->string('country_code', 10);
             $table->string('phone')->nullable();
-            $table->bigInteger('gender_id')->default(null);
+            $table->unsignedInteger('gender_id')->default(null);
             $table->boolean('is_active')->default(0);
             $table->boolean('is_verified')->default(0);
             $table->string('verification_code', 1000)->default(null);
